#72 add comments and refactor
