+++ backend/src/main/java/com/moodify/backend/services/socket/PartyRoomWebSocketHandler.java
@@ -15,6 +15,9 @@ import java.io.IOException;
 import java.net.URI;
 import java.util.*;
 
+/**
+ * WebSocket handler for managing party rooms.
+ */
 public class PartyRoomWebSocketHandler implements WebSocketHandler {
 
     private Map<String, PartyRoom> rooms = new HashMap<>();
