+++ frontend/src/app/collection/artist-profile/artist-profile.component.spec.ts
