+++ frontend/src/app/music-player/music-player.component.ts
@@ -11,23 +11,94 @@ import { filter } from 'rxjs';
     styleUrls: ['./music-player.component.css'],
 })
 export class MusicPlayerComponent {
+    /**
+     * The current track being played.
+     */
     currentTrack!: TrackTO;
+
+    /**
+     * The HTML audio element used for playing the track.
+     */
     audio!: HTMLAudioElement;
+
+    /**
+     * Indicates whether the track is currently playing.
+     */
     isPlaying = false;
+
+    /**
+     * Indicates whether shuffle mode is enabled.
+     */
     isShuffle = false;
+
+    /**
+     * The progress of the track playback, in percentage.
+     */
     progress = 0;
+
+    /**
+     * The volume level of the audio player.
+     */
     volume = 1;
+
+    /**
+     * Indicates whether the volume slider is visible.
+     */
     showVolumeSlider = false;
+
+    /**
+     * The position of the track in seconds.
+     */
     trackPosition = 0;
+
+    /**
+     * The icon representing the volume level.
+     */
     volumeSymbol = 'volume_up';
+
+    /**
+     * Indicates whether the volume is muted.
+     */
     volumeMute = false;
+
+    /**
+     * The current volume level before muting.
+     */
     currentVolume = 1;
+
+    /**
+     * Indicates whether repeat mode is enabled.
+     */
     repeat = false;
+
+    /**
+     * Indicates whether the current track is marked as favorite.
+     */
     isFavorite = false;
+
+    /**
+     * The source URL of the track's album cover image.
+     */
     imageSrc?: string;
+
+    /**
+     * Indicates whether the music player is visible.
+     */
     showMusicPlayer = true;
+
+    /**
+     * Indicates whether the queue is visible.
+     */
     showQueue = false;
+
+    /**
+     * The queue of tracks to be played.
+     */
     trackQueue: TrackTO[] = [];
+
+    /**
+     * The index of the current track in the track queue.
+     */
     currentTrackIndex = 0;
 
     constructor(
@@ -61,6 +132,9 @@ export class MusicPlayerComponent {
             });
     }
 
+    /**
+     * Sets up the audio player for the current track.
+     */
     setupAudioPlayer(): void {
         if (this.audio) {
             this.resetAudioPlayer();
@@ -71,6 +145,9 @@ export class MusicPlayerComponent {
         this.audio.addEventListener('timeupdate', this.updateProgress.bind(this));
     }
 
+    /**
+     * Resets the audio player to its initial state.
+     */
     resetAudioPlayer(): void {
         this.audio.src = this.currentTrack.preview;
         this.audio.removeEventListener('timeupdate', this.updateProgress.bind(this));
@@ -78,6 +155,9 @@ export class MusicPlayerComponent {
         this.isPlaying = false;
     }
 
+    /**
+     * Updates the progress of the track playback.
+     */
     updateProgress(): void {
         this.progress = (this.audio.currentTime / this.audio.duration) * 100;
 
@@ -86,42 +166,69 @@ export class MusicPlayerComponent {
         }
     }
 
+    /**
+     * Repeats the current track from the beginning.
+     */
     repeatTrack(): void {
         this.audio.currentTime = 0;
         this.audio.play();
     }
 
+    /**
+     * Toggles the play/pause state of the audio player.
+     */
     togglePlay() {
         this.isPlaying ? this.audio.pause() : this.audio.play();
         this.isPlaying = !this.isPlaying;
     }
 
+    /**
+     * Toggles shuffle mode.
+     */
     toggleShuffle() {
         this.isShuffle = !this.isShuffle;
         this.musicPlayerService.toggleShuffle();
     }
 
+    /**
+     * Toggles repeat mode.
+     */
     toggleRepeat() {
         this.repeat = !this.repeat;
     }
 
+    /**
+     * Skips to the next track in the queue.
+     */
     skip() {
         this.musicPlayerService.getNextTrack();
     }
 
+    /**
+     * Skips to the previous track in the queue.
+     */
     previous() {
         this.musicPlayerService.getPreviousTrack();
     }
 
+    /**
+     * Toggles the visibility of the volume slider.
+     */
     toggleVolumeSlider() {
         this.showVolumeSlider = !this.showVolumeSlider;
     }
 
+    /**
+     * Adjusts the position of the track based on the track position slider.
+     */
     adjustPosition() {
         const currentAdjustment = (this.trackPosition * this.audio.duration) / 100;
         this.audio.currentTime = currentAdjustment;
     }
 
+    /**
+     * Toggles the mute state of the audio player.
+     */
     toggleMute() {
         if (!this.volumeMute) {
             this.currentVolume = this.audio.volume;
@@ -129,6 +236,10 @@ export class MusicPlayerComponent {
         this.volumeMute ? (this.audio.volume = this.currentVolume) : (this.audio.volume = 0);
         this.volumeMute = !this.volumeMute;
     }
+
+    /**
+     * Likes the current track.
+     */
     likeSong(): void {
         if (!this.isFavorite) {
             this.backendCommunicationService.addToLikedTracks(this.currentTrack).subscribe(
@@ -158,6 +269,9 @@ export class MusicPlayerComponent {
         }
     }
 
+    /**
+     * Checks if the current track is liked by the user.
+     */
     checkIfLiked(): void {
         if (this.backendCommunicationService.userProfile == undefined || this.isFavorite) return;
         this.isFavorite = this.backendCommunicationService.userProfile.personalLibrary.likedTracks.some((track) => {
